+++ def/def/defiNet.hpp
@@ -1,6 +1,6 @@
 // *****************************************************************************
 // *****************************************************************************
-// Copyright 2013, Cadence Design Systems
+// Copyright 2013 - 2017, Cadence Design Systems
 // 
 // This  file  is  part  of  the  Cadence  LEF/DEF  Open   Source
 // Distribution,  Product Version 5.8. 
@@ -20,9 +20,9 @@
 // For updates, support, or to become part of the LEF/DEF Community,
 // check www.openeda.org for details.
 // 
-//  $Author: dell $
-//  $Revision: #1 $
-//  $Date: 2017/06/06 $
+//  $Author: icftcm $
+//  $Revision: #2 $
+//  $Date: 2017/06/19 $
 //  $State:  $
 // *****************************************************************************
 // *****************************************************************************
@@ -67,13 +67,13 @@ public:
   defiPath*   path(int index);
   const defiPath*   path(int index) const;
 
-  void bumpPaths(int size);
+  void bumpPaths(long long size);
 
 protected:
   char*      type_;
   char*      wireShieldName_;    // It only set from specialnet SHIELD, 5.4
   int        numPaths_;
-  int        pathsAllocated_;
+  long long  pathsAllocated_;
   defiPath** paths_;
 
   defrData  *defData;
@@ -130,33 +130,33 @@ public:
   defiWire*   wire(int index);
   const defiWire*   wire(int index) const;
 
-  void bumpName(int size);
-  void bumpPins(int size);
-  void bumpPaths(int size);
+  void bumpName(long long size);
+  void bumpPins(long long  size);
+  void bumpPaths(long long  size);
   void clear();
 
 protected:
-  char* name_;            // name.
-  int nameSize_;          // allocated size of name.
-  int numPins_;           // number of pins used in array.
-  int pinsAllocated_;     // number of pins allocated in array.
-  char** instances_;      // instance names for connections
-  char** pins_;           // pin names for connections
-  char* synthesized_;     // synthesized flags for pins
-  char* musts_;           // must-join flags
+  char*         name_;            // name.
+  int           nameSize_;          // allocated size of name.
+  int           numPins_;           // number of pins used in array.
+  long long     pinsAllocated_;     // number of pins allocated in array.
+  char**        instances_;      // instance names for connections
+  char**        pins_;           // pin names for connections
+  char*         synthesized_;     // synthesized flags for pins
+  char*         musts_;           // must-join flags
 
   // WMD -- the following will be removed by the next release
-  char isFixed_;        // net type
-  char isRouted_;
-  char isCover_;
-  defiPath** paths_;   // paths for this subnet
-  int numPaths_;       // number of paths used
-  int pathsAllocated_; // allocated size of paths array
+  char       isFixed_;        // net type
+  char       isRouted_;
+  char       isCover_;
+  defiPath** paths_;          // paths for this subnet
+  int        numPaths_;       // number of paths used
+  long long  pathsAllocated_; // allocated size of paths array
 
-  int numWires_;          // number of wires defined in the subnet
-  int wiresAllocated_;    // number of wires allocated in the subnet
-  defiWire** wires_;      // this replace the paths
-  char* nonDefaultRule_;
+  int        numWires_;          // number of wires defined in the subnet
+  long long  wiresAllocated_;    // number of wires allocated in the subnet
+  defiWire** wires_;             // this replace the paths
+  char*      nonDefaultRule_;
 
   defrData *defData;
 };
@@ -222,12 +222,12 @@ public:
   defiPath*         path(int index);
   const defiPath*   path(int index) const;
 
-  void bumpPaths(int size);
+  void bumpPaths(long long size);
 
 protected:
   char*      name_;
   int        numPaths_;
-  int        pathsAllocated_;
+  long long  pathsAllocated_;
   defiPath** paths_;
 
   defrData *defData;
@@ -433,12 +433,12 @@ public:
   void print(FILE* f) const;
 
 
-  void bumpName(int size);
-  void bumpPins(int size);
-  void bumpProps(int size);
-  void bumpSubnets(int size);
-  void bumpPaths(int size);
-  void bumpShieldNets(int size);
+  void bumpName(long long size);
+  void bumpPins(long long size);
+  void bumpProps(long long size);
+  void bumpSubnets(long long size);
+  void bumpPaths(long long  size);
+  void bumpShieldNets(long long size);
 
   // The method freeWire() is added is user select to have a callback
   // per wire within a net This is an internal method and is not public
@@ -451,16 +451,16 @@ public:
   void clearVia();
 
 protected:
-  char* name_;          // name.
-  int nameSize_;        // allocated size of name.
-  int numPins_;         // number of pins used in array.
-  int pinsAllocated_;   // number of pins allocated in array.
-  char** instances_;    // instance names for connections
-  char** pins_;         // pin names for connections
-  char* musts_;         // must-join flags for pins
-  char* synthesized_;   // synthesized flags for pins
-  int weight_;          // net weight
-  char hasWeight_;      // flag for optional weight
+  char*     name_;          // name.
+  int       nameSize_;      // allocated size of name.
+  int       numPins_;       // number of pins used in array.
+  long long pinsAllocated_; // number of pins allocated in array.
+  char**    instances_;     // instance names for connections
+  char**    pins_;          // pin names for connections
+  char*     musts_;         // must-join flags for pins
+  char*     synthesized_;   // synthesized flags for pins
+  int       weight_;        // net weight
+  char      hasWeight_;     // flag for optional weight
 
   // WMD -- the following will be removed by the nex release
   char isFixed_;        // net type
@@ -475,14 +475,15 @@ protected:
   char**  propValues_;  // Prop values All in strings!
   double* propDValues_; // Prop values in numbers!
   char*   propTypes_;   // Prop types, 'I' - Integer, 'R' - Real, 'S' - String
-  int propsAllocated_;  // allocated size of props array
-  int numSubnets_;      // num of subnets in array
-  defiSubnet** subnets_; // Prop names
-  int subnetsAllocated_; // allocated size of props array
-  double cap_;           // cap value
-  char* source_;
-  int    fixedbump_;     // 5.4.1
-  double frequency_;     // 5.4.1
+
+  long long    propsAllocated_;   // allocated size of props array
+  int          numSubnets_;       // num of subnets in array
+  defiSubnet** subnets_;          // Prop names
+  long long    subnetsAllocated_; // allocated size of props array
+  double       cap_;              // cap value
+  char*        source_;
+  int          fixedbump_;     // 5.4.1
+  double       frequency_;     // 5.4.1
   char* pattern_;
   char* original_;
   char* use_;
@@ -490,54 +491,54 @@ protected:
   int   style_;
 
   // WMD -- the following will be removed by the nex release
-  defiPath** paths_;   // paths for this subnet
-  int numPaths_;       // number of paths used
-  int pathsAllocated_; // allocated size of paths array
+  defiPath** paths_;          // paths for this subnet
+  int        numPaths_;       // number of paths used
+  long long  pathsAllocated_; // allocated size of paths array
 
   double voltage_;
 
-  int numWires_;         // number of wires defined in the net
-  int wiresAllocated_;   // allocated size of wire paths array
-  defiWire** wires_;     // this replace the paths
+  int         numWires_;         // number of wires defined in the net
+  long long   wiresAllocated_;   // allocated size of wire paths array
+  defiWire**  wires_;            // this replace the paths
 
-  int widthsAllocated_;
-  int numWidths_;
-  char** wlayers_;
-  double* wdist_;
+  long long   widthsAllocated_;
+  int         numWidths_;
+  char**      wlayers_;
+  double*     wdist_;
 
-  int spacingAllocated_;
-  int numSpacing_;
-  char** slayers_;
-  double* sdist_;
-  double* sleft_;
-  double* sright_;
-  int xTalk_;
+  long long   spacingAllocated_;
+  int         numSpacing_;
+  char**      slayers_;
+  double*     sdist_;
+  double*     sleft_;
+  double*     sright_;
+  int         xTalk_;
 
-  int numVpins_;
-  int vpinsAllocated_;
-  defiVpin** vpins_;
+  int         numVpins_;
+  long long   vpinsAllocated_;
+  defiVpin**  vpins_;
 
-  int numShields_;            // number of SHIELD paths used
-  int shieldsAllocated_;      // allocated size of SHIELD paths array
-  defiShield** shields_;      // SHIELD data 
-  int numNoShields_;          // number of NOSHIELD paths used
+  int          numShields_;            // number of SHIELD paths used
+  long long    shieldsAllocated_;      // allocated size of SHIELD paths array
+  defiShield** shields_;               // SHIELD data 
+  int          numNoShields_;          // number of NOSHIELD paths used
 
-  int numShieldNet_;          // number of SHIELDNETS used in array.
-  int shieldNetsAllocated_;   // number of SHIELDNETS allocated in array.
-  char** shieldNet_;          // name of the SHIELDNET
+  int          numShieldNet_;          // number of SHIELDNETS used in array.
+  long long    shieldNetsAllocated_;   // number of SHIELDNETS allocated in array.
+  char**       shieldNet_;             // name of the SHIELDNET
 
-  int numPolys_;                    // 5.6
-  char** polygonNames_;             // 5.6 layerName for POLYGON
-  int polysAllocated_;              // 5.6
-  struct defiPoints** polygons_;    // 5.6
-  int*   polyMasks_;
+  int          numPolys_;              // 5.6
+  char**       polygonNames_;          // 5.6 layerName for POLYGON
+  long long    polysAllocated_;        // 5.6
+  struct defiPoints** polygons_;       // 5.6
+  int*         polyMasks_;
   char** polyRouteStatus_;
   char** polyShapeTypes_;
   char** polyRouteStatusShieldNames_;
 
-  int numRects_;                    // 5.6
-  int rectsAllocated_;              // 5.6
-  char** rectNames_;                // 5.6
+  int        numRects_;                    // 5.6
+  long long  rectsAllocated_;              // 5.6
+  char**     rectNames_;                   // 5.6
   int* xl_;
   int* yl_;
   int* xh_;
@@ -549,18 +550,17 @@ protected:
   
 
   struct defiPoints** viaPts_;      // 5.8                  
-  char** viaNames_;                   
-  int   numPts_;                    
-  int   ptsAllocated_;
-  int*  viaOrients_;
-  int*  viaMasks_;
-  char** viaRouteStatus_;
-  char** viaRouteStatusShieldNames_;
-  char** viaShapeTypes_;
+  char**              viaNames_;                   
+  int                 numPts_;                    
+  long long           ptsAllocated_;
+  int*                viaOrients_;
+  int*                viaMasks_;
+  char**              viaRouteStatus_;
+  char**              viaRouteStatusShieldNames_;
+  char**              viaShapeTypes_;
 
   defrData *defData;
 };
-
 
 END_LEFDEF_PARSER_NAMESPACE
